bug: 인트형이 아니라 롱이였네 id 이거 바꿈

+++ TimeManager/Forms/AddScheduleForm.cs
@@ -20,14 +20,14 @@ namespace TimeManager.Forms
 
         DayOfWeek dayOfWeek;
 
-        int lastID;
+        long lastID;
 
         List<WeeklyDateTimeBlock> weeklyDateTimeBlocks;
         WeeklyDateTimeBlock[] weeklyBlock = new WeeklyDateTimeBlock[7];
 
         bool[] daysBool = new bool[] { false, false, false, false, false, false, false };
 
-        public AddScheduleForm(MainForm mainForm, int lastID)
+        public AddScheduleForm(MainForm mainForm, long lastID)
         {
             InitializeComponent();
